dual-layer/requirements.txt
neo4j==5.23.0
pyyaml==6.0.1
pydantic==2.7.0
requests==2.31.0
tenacity==8.2.3
tqdm==4.66.2
numpy==1.26.4
scikit-learn==1.4.2
rapidfuzz==3.9.0

# PDF processing
PyMuPDF==1.23.14

# Hugging Face dependencies
transformers==4.36.2
torch==2.1.2
accelerate==0.25.0
datasets==2.16.1

# GUI dependencies
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.4
